Added force drop database fn impl

+++ sqlx-cli/src/database.rs
@@ -33,7 +33,11 @@ pub async fn drop(connect_opts: &ConnectOpts, confirm: bool, force: bool) -> any
     let exists = crate::retry_connect_errors(connect_opts, Any::database_exists).await?;
 
     if exists {
-        Any::drop_database(&connect_opts.database_url).await?;
+        if force {
+            Any::force_drop_database(&connect_opts.database_url).await?;
+        } else {
+            Any::drop_database(&connect_opts.database_url).await?;
+        }
     }
 
     Ok(())
